requirements.txt
Pillow~=8.4.0
opencv-python~=4.6.0.66